Rename 'cmd_server' to 'server' in cli commands.
